+++ lib/shakapacker/compiler_strategy.rb
@@ -3,14 +3,14 @@ require_relative "digest_strategy"
 
 module Shakapacker
   class CompilerStrategy
-    def self.from_config
-      strategy_from_config = Shakapacker.config.compiler_strategy
+    def self.from_config(instance)
+      strategy_from_config = instance.config.compiler_strategy
 
       case strategy_from_config
       when "mtime"
-        Shakapacker::MtimeStrategy.new
+        Shakapacker::MtimeStrategy.new(instance)
       when "digest"
-        Shakapacker::DigestStrategy.new
+        Shakapacker::DigestStrategy.new(instance)
       else
         raise "Unknown strategy '#{strategy_from_config}'. " \
               "Available options are 'mtime' and 'digest'."